Extract delivery profile detail cards into their own components
The profile detail route had grown to 1150 lines holding fifteen nested
components. Each card and section now lives beside the other shipping
components, and the method sheet's URL state reaches them as a prop
rather than through a route import.

No behaviour change.
